Support multi routine Dispatch
A mutex was added to prevent data races.

Perhaps a closer look at wl_display_dispatch_pending and
https://wayland.freedesktop.org/docs/html/apb.html could result in
an implementation that does not require users of go-wayland to
remember locking and unlocking. While this might be possible, it
could prove difficult and require significant rewriting.
It might also be prevented due to lack of non-blocking conn.Read in Go.