Rework multi-goroutine dispatching
The previous implementation has data races.
One example is that IdleNotification.Close calls Context.Unregister
which races with Context.Dispatch.

The new approach adds GetDispatch that can be called safely in another
goroutine and returns the dispatch function which should be called on
the same goroutine as the rest of the Context.

This allows waiting for a new message to dispatch in another goroutine,
which solves the blocking, and allows calling the dispatch function in
the main goroutine, which solves the data races.

The Dispatch function remains for use cases where blocking, single
goroutine, dispatch is sufficient. It effectively reverts back to
before the previous multi goroutine implementation.